src/test/java/com/zepben/ewbnetworkserver/patch/PatchFeatureTest.java
/*
 * Copyright 2020 Zeppelin Bend Pty Ltd
 * This file is part of ewb-network-server.
 *
 * ewb-network-server is free software: you can redistribute it and/or modify
 * it under the terms of the GNU Affero General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * ewb-network-server is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU Affero General Public License for more details.
 *
 * You should have received a copy of the GNU Affero General Public License
 * along with ewb-network-server.  If not, see <https://www.gnu.org/licenses/>.
 */

package com.zepben.ewbnetworkserver.patch;

import com.zepben.testutils.junit.SystemLogExtension;
import com.zepben.testutils.mockito.DefaultAnswer;
import io.vertx.core.json.JsonObject;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.RegisterExtension;

import java.util.Optional;

import static org.hamcrest.MatcherAssert.assertThat;
import static org.hamcrest.Matchers.equalTo;
import static org.mockito.Mockito.mock;

public class PatchFeatureTest {

    @RegisterExtension
    public SystemLogExtension systemOutRule = SystemLogExtension.SYSTEM_OUT.captureLog().muteOnSuccess();

    private final PatchResult patchResult = new PatchResult(1);
    private final Patch patch = mock(Patch.class, DefaultAnswer.of(PatchResult.class, patchResult));

    @Test
    public void detectsMalformedPatchFeatures() {
        assertThat(PatchFeature.parse(patch, "something", new JsonObject().put("key", "value")), equalTo(Optional.empty()));
        assertThat(patchResult.errors().get(patchResult.errors().size() - 1), equalTo("Ignoring malformed something feature: No value found for required key 'type'. `{\"key\":\"value\"}`"));
    }

}
